Finalize project
- bump version to 0.8.0
- switch to using rustls with webpki which doesn't require
any dynamic linking, which is challenging when cross-compiling
- Update the README.md with cross compiling instructions
- Add new line after saving a data point for all 3 tasks.
This makes it easier to parse back the data and it is only a
1 character addition
- Pop values from MovingAverage VecDeque since it does
not happen automatically
- Refactor parts of main into lib functions to reduce
complexity, more can be done but it is ok.
- Loop task 1. The while let loop does not run infinitely
otherwise, since it return Ok(()) after a value has been
successfully sent and written. Thus the while let isn't
necessary any more and can be a simple if let.
- Time the computation -> file write of task2 and task3,
and write the delays to each file separately.